refactor(lib): create `lib/discord` subfolder
